pyproject.toml
[tool.poetry]
name = "hohotach"
version = "0.0.1"
description = "unpowerful tool to ask grok via vim"
authors = [
    "Banakh Andrei <uuetsukeu@mail.ru>"
]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.9"
telethon = "^1.40.0"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
launch = "app.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
